+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v1.4.0 (2017-08-XX)
+## v1.4.0 (2017-08-20)
 - added command ntp to check NTP status (#18)
 - merged check_threshold and get_perfdata into one function: check_threshold_and_get_perfdata
 - added command hastatus to check the status of an high availability pair (#25)